Remove 68000 library multiplies from the blit, map and hash paths
Follow-up to the cleanup pass, from findings that arrived after it.
Every item here was confirmed in the disassembly before and after, not
inferred from the source.

A 32-bit multiply by a variable is a __mulsi3 call on the 68000, a few
hundred cycles. Three places were paying for one where both operands
fit in 16 bits, so a bare MULU.W does the job:

- id_vl_stdl.c: the row offset y * stride in ten blit and fill paths,
  including the aligned fast paths added for the tile blitter, which
  were calling out to the library twice per blit. The eleventh site,
  the cold non-drift fallback in SurfaceToSelf, keeps the signed form:
  unlike the others it has no clamp making y non-negative, so the
  unsigned cast would be unsafe there for no gain.

- id_ca.c: the y * width in CA_TileAtPos, CA_TilePtrAtPos and
  CA_SetTileAtPos, which the physics and collision code calls
  constantly.

- id_str.c: the hash finaliser was hash * 0x9E3779B1, a __mulsi3 on
  every string-table lookup. Folding the high half down with a shift
  gives the same mixing and compiles to SWAP. The per-character byte
  also went through a sign extension to long; making it unsigned char
  drops that too.

Measured on a plain STE in Hatari, matched demo scenes: a name lookup
falls from 0.277ms to 0.233ms, and the frame is 1.2-2.2% shorter. The
lookup win also applies to level loading, which does far more of them.

Also adds a profile-only counter for CK_INT/CK_ACTION name lookups
with a calibration loop, since the 200Hz clock cannot time one call.
It reports 11-21 lookups per in-level frame, about 3.5ms - real, but
not the explanation for the game-logic bucket it was suspected to be.

+++ src/id_ca.c
@@ -1974,7 +1974,7 @@ void CA_LoadAllSounds(void)
 
 uint16_t *CA_TilePtrAtPos(int16_t x, int16_t y, int16_t plane)
 {
-	return &CA_mapPlanes[plane][y * CA_MapHeaders[ca_mapOn]->width + x];
+	return &CA_mapPlanes[plane][(uint32_t)((uint16_t)y * (uint16_t)CA_MapHeaders[ca_mapOn]->width) + x];
 }
 
 uint16_t CA_TileAtPos(int16_t x, int16_t y, int16_t plane)
@@ -1984,13 +1984,13 @@ uint16_t CA_TileAtPos(int16_t x, int16_t y, int16_t plane)
 	// (not reproduced in the exact same manner)
 	CA_MapHeader *mapheader = CA_MapHeaders[ca_mapOn];
 	if ((x >= 0) && (x < mapheader->width) && (y >= 0) && (y < mapheader->height))
-		return CA_mapPlanes[plane][y * mapheader->width + x];
+		return CA_mapPlanes[plane][(uint32_t)((uint16_t)y * (uint16_t)mapheader->width) + x];
 	return (y * x * y * x) % (ca_gfxInfoE.numTiles16 * 2 + ca_gfxInfoE.numTiles16m * 6);
 }
 
 void CA_SetTileAtPos(int16_t x, int16_t y, int16_t plane, uint16_t value)
 {
-	CA_mapPlanes[plane][y * CA_MapHeaders[ca_mapOn]->width + x] = value;
+	CA_mapPlanes[plane][(uint32_t)((uint16_t)y * (uint16_t)CA_MapHeaders[ca_mapOn]->width) + x] = value;
 }
 
 uint16_t CA_GetMapWidth()
